fix: improve state management of tab bar (#383)
* fix: improve state management of tab bar

`SBBTabBar` has two constructors:

1. One that takes a controller (which is probably kept in the caller's state).
2. One that takes a list of items and creates a controller on the fly.

The former is without major problems, but the latter has the issue that the controller is prone to changing, and thus violating some basic preconditions of the widget.

This commit tries to mitigate this issue by managing the controller in its own state, and, when the items change, recreating the controller and rebinding the ticker.

* refactor: incorporate review feedback